Changed html title from Scrumptious -> Crumbs Reorganized some components into subdirectories
