+++ public/index.html
@@ -20,7 +20,7 @@
       integrity="sha384-B0vP5xmATw1+K9KRQjQERJvTumQW0nPEzvF6L/Z6nronJ3oUOFUFpCjEUQouq2+l"
       crossorigin="anonymous"
     />
-    <title>Scrumptious</title>
+    <title>Crumbs</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
